LayoutTests/inspector/debugger/dynamic-scripts-expected.txt

CONSOLE MESSAGE: line 30: Done.
Tests that scripts for dynamically added script elements are shown in sources panel if loaded with inspector open.

Bug 99324
Debugger was enabled.
UISourceCodes:
    debugger-test.js
    dynamic-script.js
    dynamic-scripts.html
    evalSourceURL.js
    inspector-test.js
    scriptElementContentSourceURL.js
Debugger was disabled.
